IEIBUILD-2635 Convert tag refs to short names
The Task Bot determines its behaviour based on the names of the refs
that are involved in a pull request. It has known to shorten the full
ref names of branches, but not tags. By remedying this, we can
assess a pull request from a tagged team build in the same way we
would one from the tip of a team branch.